+++ packages/template-retail-react-app/app/components/passwordless-login/index.jsx
@@ -54,7 +54,7 @@ const PasswordlessLogin = ({
                         isLoading={form.formState.isSubmitting}
                     >
                         <FormattedMessage
-                            defaultMessage="Continue Securely"
+                            defaultMessage="Continue"
                             id="login_form.button.continue_securely"
                         />
                     </Button>