+++ packages/template-retail-react-app/app/pages/reset-password/index.test.jsx
@@ -13,6 +13,7 @@ import {
 } from '@salesforce/retail-react-app/app/utils/test-utils'
 import ResetPassword from '.'
 import mockConfig from '@salesforce/retail-react-app/config/mocks/default'
+import {TOO_MANY_PASSWORD_RESET_ATTEMPTS_ERROR_MESSAGE} from '@salesforce/retail-react-app/app/utils/auth-utils'
 
 const mockUseRouteMatch = jest.fn(() => ({path: '/'}))
 
@@ -88,6 +89,38 @@ test('Allows customer to generate password token', async () => {
         expect(window.location.pathname).toBe('/uk/en-GB/login')
     })
 })
+
+test.each([
+    ['no callback_uri is registered for client', 'This feature is not currently available.'],
+    [
+        'Too many password reset requests were made. Please try again later.',
+        'You reached the limit for password resets. For your security, wait 10 minutes and try again.'
+    ],
+    ['unexpected error message', 'Something went wrong. Try again!']
+])(
+    'displays correct error message when password reset fails with "%s"',
+    async (apiErrorMessage, expectedMessage) => {
+        global.server.use(
+            rest.post('*/oauth2/password/reset', (req, res, ctx) =>
+                res(ctx.delay(0), ctx.status(400), ctx.json({message: apiErrorMessage}))
+            )
+        )
+        // render our test component
+        const {user} = renderWithProviders(<MockedComponent />, {
+            wrapperProps: {siteAlias: 'uk', appConfig: mockConfig.app}
+        })
+
+        // enter credentials and submit
+        await user.type(await screen.findByLabelText('Email'), 'foo@test.com')
+        await user.click(
+            within(await screen.findByTestId('sf-auth-modal-form')).getByText(/reset password/i)
+        )
+
+        await waitFor(() => {
+            expect(screen.getByText(expectedMessage)).toBeInTheDocument()
+        })
+    }
+)
 
 test.each([
     ['base path', '/reset-password-landing'],