+++ packages/pwa-kit-create-app/CHANGELOG.md
@@ -1,4 +1,7 @@
-## v3.16.0-dev (Dec 17, 2025)
+
+## v3.15.0-dev (Nov 05, 2025)
+- Support email mode by default for passwordless login and password reset in a generated app. [#3525](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/3525)
+
 ## v3.15.0 (Dec 17, 2025)
 - Add new Google Cloud API configuration and Bonus Product configuration [#3523](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/3523)